rename metric package to telemetry
